fix: ProfileRepository.unfollow() の try catch を最小に変更

+++ src/main/kotlin/com/example/realworldkotlinspringbootjdbc/infra/ProfileRepositoryImpl.kt
@@ -178,28 +178,49 @@ class ProfileRepositoryImpl(val namedParameterJdbcTemplate: NamedParameterJdbcTe
     }
 
     override fun unfollow(username: Username, currentUserId: UserId): Either<ProfileRepository.UnfollowError, Profile> {
-        val selectUserSql = """
-            SELECT
-                users.username
-                , profiles.bio
-                , profiles.image
-                , CASE WHEN followings.id IS NOT NULL THEN 1 ELSE 0 END AS following_flg
-            FROM
-                users
-            JOIN
-                profiles
-            ON
-                users.id = profiles.user_id
-                AND users.username = :username
-            LEFT OUTER JOIN
-                followings
-            ON
-                followings.following_id = users.id
-                AND followings.follower_id = :current_user_id
-            ;
-        """.trimIndent()
-        val sqlSelectUserSqlParams = MapSqlParameterSource().addValue("username", username.value)
-            .addValue("current_user_id", currentUserId.value)
+        val profileFromDb: MutableList<MutableMap<String, Any>>
+
+        /**
+         * username を取得
+         */
+        try {
+            val selectUserSql = """
+                SELECT
+                    users.username
+                    , profiles.bio
+                    , profiles.image
+                    , CASE WHEN followings.id IS NOT NULL THEN 1 ELSE 0 END AS following_flg
+                FROM
+                    users
+                JOIN
+                    profiles
+                ON
+                    users.id = profiles.user_id
+                    AND users.username = :username
+                LEFT OUTER JOIN
+                    followings
+                ON
+                    followings.following_id = users.id
+                    AND followings.follower_id = :current_user_id
+                ;
+            """.trimIndent()
+            val sqlSelectUserSqlParams = MapSqlParameterSource().addValue("username", username.value)
+                .addValue("current_user_id", currentUserId.value)
+            profileFromDb = namedParameterJdbcTemplate.queryForList(selectUserSql, sqlSelectUserSqlParams)
+        } catch (e: Throwable) {
+            return ProfileRepository.UnfollowError.Unexpected(e, username, currentUserId).left()
+        }
+
+        /**
+         * user が存在しなかった時 NotFoundError
+         */
+        if (profileFromDb.isEmpty()) {
+            return ProfileRepository.UnfollowError.NotFoundProfileByUsername(username, currentUserId).left()
+        }
+
+        /**
+         * user が存在した時 アンフォロー
+         */
         val deleteFollowingsSql = """
             DELETE FROM
                 followings
@@ -215,27 +236,15 @@ class ProfileRepositoryImpl(val namedParameterJdbcTemplate: NamedParameterJdbcTe
             .addValue("username", username.value)
             .addValue("current_user_id", currentUserId.value)
         return try {
-            val profileFromDb = namedParameterJdbcTemplate.queryForList(selectUserSql, sqlSelectUserSqlParams)
-            when (profileFromDb.isEmpty()) {
-                /**
-                 * プロフィールが存在しなかったら、UnfollowError.NotFoundProfileByUsername
-                 */
-                true -> ProfileRepository.UnfollowError.NotFoundProfileByUsername(username, currentUserId).left()
-                /**
-                 * プロフィールが存在したら、アンフォロー
-                 */
-                false -> {
-                    namedParameterJdbcTemplate.update(deleteFollowingsSql, deleteFollowingsSqlParams)
-                    profileFromDb.map {
-                        Profile.newWithoutValidation(
-                            Username.newWithoutValidation(it["username"].toString()),
-                            Bio.newWithoutValidation(it["bio"].toString()),
-                            Image.newWithoutValidation(it["image"].toString()),
-                            false
-                        )
-                    }[0].right()
-                }
-            }
+            namedParameterJdbcTemplate.update(deleteFollowingsSql, deleteFollowingsSqlParams)
+            profileFromDb.map {
+                Profile.newWithoutValidation(
+                    Username.newWithoutValidation(it["username"].toString()),
+                    Bio.newWithoutValidation(it["bio"].toString()),
+                    Image.newWithoutValidation(it["image"].toString()),
+                    false
+                )
+            }[0].right()
         } catch (e: Throwable) {
             ProfileRepository.UnfollowError.Unexpected(e, username, currentUserId).left()
         }
